Undeprecated interface configuration. Added more Define methods.
I've revised my thinking about interface configuration.  I think its a
good option and should be supported.  Field scope is now ignored across
the board.  If the developer does not want to map a field and is using
the interface configuration approach, then the field should be annotated
with IQIgnore.